Bug fix in oauth browser flow
I noticed that I was unable to use the popup oauth browser flow, and the
cause seems to be because the state is never passed to
`BaseBrowserFlow.authorizeUrl`. Since this is stored in `me` right before
calling `BaseBrowserFlow.authorizeUrl`, I assume this isn't meant as a
parameter and instead should be read from `this` (as the other options are).